Added the ability to proxy the WASI syscalls for extendability reasons
